Fix 'undefined reference' errors, makedeps, docs
* The 'undefined reference' linking errors were due to malformed
libLLVM.so, which, in turn, got screwed up because of a failure in
tools/llvm-shlib/CMakeLists.txt, which, in turn, was caused by the
BUILD_SHARED_LIBS flag to Cmake (symbols fail to export). In fact,
we don't even need the BUILD_SHARED_LIBS flag. This fixes issue #1.

* Cmake is not part of base-devel, so we need to explicitly add it.
Also, we don't use chrpath anymore (need to double-check if that's OK).

* Clean up the documentation a little bit.