Fix (some of) the symbols exported from libLLVM.so
This is another attempt (but still not enough) to solve the issues with
builds against the libLLVM.so shared lib (e.g. Mesa). The problem is
that some symbols are not exported as expected. What probably remains
only is the bug here: https://llvm.org/bugs/show_bug.cgi?id=24157