Move the OCaml bindings check back to build()
The most time consuming operation is preparing the working copies, and
by checking in prepare() we still can't spare that time for the users.
However, checking in prepare() may break unnecessarily `makepkg -o'.
Finally, even just logic calls for this check to be performed in
build(), where it also did originally belong. This partially reverts
commit 11fbb41699576e0c38ee39742008dd34fa74607f.