Fix _ocamlver() for Pacman 5.1.0
Makepkg apparently now parses the PKGBUILDS in advance and doesn't like
seeing empty pkgvers for dependencies. With our split package for the
OCaml bindings we declare a dependency on the installed OCaml version,
so that upgrades to OCaml wouldn't break the bindings. When llvm-ocaml
is built, OCaml is installed and thus _ocamlver() returns a valid
version. But now we also need to return a valid version before OCaml is
installed. It doesn't need to be meaningful, as it'll get updated again
to the real version if package_llvm-ocaml-svn() gets to be run later.